requirements.txt
Flask==3.1.2
flask-cors==6.0.2
Pillow>=12.1.1
